backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-generativeai==0.8.0
python-dotenv==1.0.1
pydantic==2.9.0
slowapi==0.1.9
